Create Code
This is a convolutional autoencoder designed for CIFAR10. The training portion is presented here. This code is inspired from https://gist.github.com/okiriza/16ec1f29f5dd7b6d822a0a3f2af39274, which was written for MNIST dataset. This code is designed for CIFAR10 and it uses MaxUnpool2d in the decoder section.